Convert Preferences to Route-based component
