+++ src/renderer/components/+preferences/extensions.tsx
@@ -19,36 +19,20 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import type { RouteProps } from "react-router";
-import { buildURL } from "../utils/buildUrl";
+import { observer } from "mobx-react";
+import React from "react";
+import { AppPreferenceRegistry } from "../../../extensions/registries";
+import { ExtensionSettings } from "./preferences";
 
-export const preferencesRoute: RouteProps = {
-  path: "/preferences"
-};
+export const Telemetry = observer(() => {
+  const extensions = AppPreferenceRegistry.getInstance().getItems();
 
-export const appRoute: RouteProps = {
-  path: `${preferencesRoute.path}/app`
-};
-
-export const proxyRoute: RouteProps = {
-  path: `${preferencesRoute.path}/proxy`
-};
-
-export const kubernetesRoute: RouteProps = {
-  path: `${preferencesRoute.path}/kubernetes`
-};
-
-export const telemetryRoute: RouteProps = {
-  path: `${preferencesRoute.path}/telemetry`
-};
-
-export const extensionRoute: RouteProps = {
-  path: `${preferencesRoute.path}/extensions`
-};
-
-export const preferencesURL = buildURL(preferencesRoute.path);
-export const appURL = buildURL(appRoute.path);
-export const proxyURL = buildURL(proxyRoute.path);
-export const kubernetesURL = buildURL(kubernetesRoute.path);
-export const telemetryURL = buildURL(telemetryRoute.path);
-export const extensionURL = buildURL(extensionRoute.path);
+  return (
+    <section id="extensions">
+      <h2>Extensions</h2>
+      {extensions.filter(e => !e.showInPreferencesTab).map((extension) =>
+        <ExtensionSettings key={extension.id} {...extension}/>
+      )}
+    </section>
+  );
+});
